Move `Config` into `search`
