+++ monument/lib/src/lib.rs
@@ -14,42 +14,5 @@ mod utils;
 
 pub use composition::Composition;
 pub use error::{Error, Result};
-pub use search::{Progress, SearchData, SearchUpdate};
+pub use search::{Config, Progress, SearchData, SearchUpdate};
 pub use utils::OptionalRangeInclusive;
-
-use query::{CallIdx, MethodIdx, MethodVec, Query};
-
-/// Configuration parameters for Monument which **don't** change which compositions are emitted.
-#[derive(Debug, Clone)]
-pub struct Config {
-    /* General */
-    /// Number of threads used to generate compositions.  If `None`, this uses the number of
-    /// **physical** CPU cores (i.e. ignoring hyper-threading).
-    pub thread_limit: Option<usize>,
-
-    /* Graph Generation */
-    /// The maximum graph size, in chunks.  If a search would produce a graph bigger than this, it
-    /// is aborted.
-    pub graph_size_limit: usize,
-
-    /* Search */
-    pub queue_limit: usize,
-    /// If `true`, the data structures used by searches will be leaked using [`std::mem::forget`].
-    /// This massively improves the termination speed (because all individual allocations don't
-    /// need to be freed), but only makes sense for the CLI, where Monument will do exactly one
-    /// search run before terminating (thus returning the memory to the OS anyway).
-    pub leak_search_memory: bool,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            thread_limit: None,
-
-            graph_size_limit: 100_000,
-
-            queue_limit: 10_000_000,
-            leak_search_memory: false,
-        }
-    }
-}
